Refactor sort pods by usage
